+++ NEWS.md
@@ -1,5 +1,8 @@
 # mapdeck 0.1.006
 
+* `mapdeck()` argument order changed so `data` is first ( to work better with pipes ) 
+* `update_view` and `focus_layer` added to focus layers on data
+* bearing and pitch maintained on data layer updates
 * `bearing` argument to `mapdeck()`
 * `add_geojson()` fully supported
 * `add_sf()` convenience function